requirements.txt
celery==5.3.6
click==8.1.7
flask==3.0.3
flask_sqlalchemy==3.1.1
flask_cors==4.0.0
langchain==0.2.14
langchain-openai==0.1.22
python-dotenv==1.0.0
Werkzeug==3.0.0
invoke==2.2.0
pinecone-client==5.0.1
watchdog==3.0.0
tiktoken==0.7.0
uuid==1.30
langfuse==1.0.18
backoff==2.2.1
redis==5.0.3
azure-storage-blob==12.19.1
langchain-community==0.2.12
langchain-pinecone==0.1.3
pypdf==3.15.4
unstructured==0.11.8
python-docx==1.1.0
docx2txt==0.8
pandas==2.2.2
google-api-python-client==2.127.0
html2text==2024.2.26
chromadb==0.5.0
certifi==2024.2.2
prophet==1.1.5
requests~=2.31.0
boto3==1.34.140
langgraph==0.2.13
langchainhub==0.1.21